feat: Dynamically set theme color and improve attendance ranks
Set the theme color dynamically based on the user's rank when on the TimeAttendance page. This enhances the mobile experience by visually reflecting the user's status.

Also, refactored the rank titles to include a specific theme color for each rank, improving visual differentiation. The Layout component now resets the theme color when not on the homepage to prevent inconsistencies.